Repackage command classes into commands package
