+++ src/main/java/seedu/duke/commands/AddActivityCommand.java
@@ -1,16 +1,17 @@
-package seedu.duke;
+package seedu.duke.commands;
 
+import seedu.duke.*;
 import seedu.duke.exceptions.FlirtForkException;
 
 
-public class AddActivityCommand extends Command{
+public class AddActivityCommand extends Command {
     private String description;
     public AddActivityCommand(String description) {
         this.description = description;
     }
 
     @Override
-    public void execute(FavouritesList favourites, FoodList foods, ActivityList activities,Ui ui,
+    public void execute(FavouritesList favourites, FoodList foods, ActivityList activities, Ui ui,
                         Storage storage, UserDetails userDetails, GiftList gifts) throws FlirtForkException {
         Activity activity = new Activity(description, "default location", "default price", "U");
         favourites.addFavourite(activity);
