+++ src/main/java/seedu/duke/commands/GenerateItineraryCommand.java
@@ -1,10 +1,11 @@
-package seedu.duke;
+package seedu.duke.commands;
 
+import seedu.duke.*;
 import seedu.duke.exceptions.FlirtForkException;
 import java.util.logging.Logger;
 import java.util.logging.Level;
 
-public class GenerateItineraryCommand extends Command{
+public class GenerateItineraryCommand extends Command {
     private static final Logger LOGGER = Logger.getLogger(GenerateItineraryCommand.class.getName());
     private String preferredLocation;
     private String preferredPrice;
