+++ src/main/java/seedu/duke/commands/GenerateSmartItineraryCommand.java
@@ -1,5 +1,6 @@
-package seedu.duke;
+package seedu.duke.commands;
 
+import seedu.duke.*;
 import seedu.duke.exceptions.FlirtForkException;
 import java.util.logging.Logger;
 import java.util.logging.Level;
@@ -9,7 +10,7 @@ import java.util.logging.Level;
  * This command considers the user's preferred location and cuisine to create
  * a personalized itinerary that includes food and activities.
  */
-public class GenerateSmartItineraryCommand extends Command{
+public class GenerateSmartItineraryCommand extends Command {
     private static final Logger LOGGER = Logger.getLogger(GenerateItineraryCommand.class.getName());
     private String preferredLocation;
     private String preferredCuisine;
