+++ src/main/java/seedu/duke/commands/ListOptionsCommand.java
@@ -1,5 +1,6 @@
-package seedu.duke;
+package seedu.duke.commands;
 
+import seedu.duke.*;
 import seedu.duke.exceptions.FlirtForkException;
 
 public class ListOptionsCommand extends Command {
